requirements/test.txt
blessings==1.6.1
bpython==0.17.1
certifi==2018.1.18
chardet==3.0.4
curtsies==0.3.0
decorator==4.2.1
greenlet==0.4.13
idna==2.6
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
multidict==4.1.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
tox==2.9.1
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
vcrpy==1.11.1
virtualenv==15.1.0
wcwidth==0.1.7
wrapt==1.10.11
yarl==1.1.1
